requirements.txt
fastapi
uvicorn
sqlalchemy
python-dotenv
requests
SpeechRecognition
gTTS
pydub
elevenlabs
httpcore>=0.15.0
httpx==0.23.0
deep-translator
joblib
numpy
scikit-learn
langchain_google_genai
langchain_core
elevenlabs
langchain_ollama
standard-aifc
python-multipar